EWSCv2.0
A image recognition script developed using OpenCV, designed for the game EVE Online created by CCP. The working principle is to capture the image of the target area, convert it into grayscale and RGB images through color concentration, and then match it with the provided template image. If the similarity exceeds 80%, it will play music. The script also monitors mouse activity. If no mouse activity is detected for two minutes after manually stopping the detection, the script will automatically restart the detection.